lite/pyproject.toml
[tool.poetry]
name = "eosce"
version = "0.1.0"
description = "Ersilia Cpound embeddings"
authors = ["Ankur Kumar <ank@leoank.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
rdkit-pypi = "2022.9.4"
click = "8.1.3"
onnxruntime = "1.14.0"
onnxruntime-gpu = {version = "1.14.0", optional = true}
joblib = "1.2.0"

[tool.poetry.extras]
gpu = ["onnxruntime-gpu"]

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
flake8 = "^5.0.4"
pyproject-flake8 = "5.0.4"
black = "^22.6.0"
darglint = "^1.8.1"

[tool.poetry.scripts]
eosce = "eosce.cli:main"

[tool.flake8]
select = "ANN,B,B9,BLK,C,D,DAR,E,F,I,S,W"
ignore = "E203,E501,W503,ANN401"
max-line-length = 88
max-complexity = 10
application-import-names = "eosce,tests"
import-order-style = "google"
docstring-convention = "google"
per-file-ignores = """
    tests/*:S101,D101,D102,D103"
    eosce/cli:DAR101
    """
[tool.black]
line-length = 88

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
